+++ src/context.js
@@ -5,6 +5,7 @@ var vars = require("./variables");
 
 export const AppContext = createContext({
     apiPath: "", setApiPath: () => { },
+    apiVersion: "", setApiVersion: () => { },
 
     apiConfig: null, setApiConfig: () => { },
     webConfig: null, setWebConfig: () => { },
@@ -44,6 +45,7 @@ export const AppContext = createContext({
 
 export const AppContextProvider = ({ children }) => {
     const [apiPath, setApiPath] = useState("");
+    const [apiVersion, setApiVersion] = useState("");
 
     const [apiConfig, setApiConfig] = useState(null);
     const [webConfig, setWebConfig] = useState(null);
@@ -230,6 +232,7 @@ export const AppContextProvider = ({ children }) => {
 
     const value = useMemo(() => ({
         apiPath, setApiPath,
+        apiVersion, setApiVersion,
 
         apiConfig, setApiConfig,
         webConfig, setWebConfig,
@@ -255,7 +258,7 @@ export const AppContextProvider = ({ children }) => {
         dlogDetailsCache, setDlogDetailsCache, loadDlogDetails,
         economyCache, setEconomyCache,
         allUsersCache, setAllUsersCache, loadAllUsers
-    }), [apiPath, apiConfig, webConfig, languages, allRoles, allPerms, allRanks, users, userProfiles, memberUIDs, curUID, curUser, curUserPerm, curUserBanner, userSettings, announcementTypes, applicationTypes, divisions, dlogDetailsCache, economyCache, allUsersCache]);
+    }), [apiPath, apiVersion, apiConfig, webConfig, languages, allRoles, allPerms, allRanks, users, userProfiles, memberUIDs, curUID, curUser, curUserPerm, curUserBanner, userSettings, announcementTypes, applicationTypes, divisions, dlogDetailsCache, economyCache, allUsersCache]);
 
     return (
         <AppContext.Provider value={value}>