+++ changelog.md
@@ -58,6 +58,7 @@
 24. Replaced `vars.ranks` with Context API `allRanks`
 25. Replaced `vars.dhpath` with Context API `apiPath`
 26. Replaced `vars.host` with `window.dhhost` (Context API was not used because this variable is mostly static and is defined locally on initial load)
+27. Replaced `vars.apiversion` with Context API `apiVersion`
 
 ## v3.3.x